+++ keba_keenergy_api/constants.py
@@ -759,6 +759,10 @@ class HeatCircuit(Enum):
         f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.heatCurveOffset",
         value_type=float,
     )
+    HEATING_CURVE_GRADIENT = EndpointProperties(
+        f"{PAYLOAD_PREFIX}.sParam.heatCircuit[%s].param.heatCurveGradient",
+        value_type=float,
+    )
 
 
 class SolarCircuit(Enum):