+++ tests/test_endpoints.py
@@ -4902,6 +4902,60 @@ class TestHeatCircuitSection:
                 ssl=False,
             )
 
+    @pytest.mark.asyncio
+    async def test_get_heating_curve_gradient(self) -> None:
+        with aioresponses() as mock_keenergy_api:
+            mock_keenergy_api.post(
+                "http://mocked-host/var/readWriteVars",
+                payload=[
+                    {
+                        "name": "APPL.CtrlAppl.sParam.heatCircuit[0].param.heatCurveGradient",
+                        "attributes": {
+                            "formatId": "fmt3p2",
+                            "longText": "Heat curve grad.",
+                            "upperLimit": "5",
+                            "lowerLimit": "0",
+                        },
+                        "value": "0.25",
+                    },
+                ],
+                headers={"Content-Type": "application/json;charset=utf-8"},
+            )
+
+            client: KebaKeEnergyAPI = KebaKeEnergyAPI(host="mocked-host")
+            data: float = await client.heat_circuit.get_heating_curve_gradient()
+
+            assert isinstance(data, float)
+            assert data == 0.25  # noqa: PLR2004
+
+            mock_keenergy_api.assert_called_once_with(
+                url="http://mocked-host/var/readWriteVars",
+                data='[{"name": "APPL.CtrlAppl.sParam.heatCircuit[0].param.heatCurveGradient", "attr": "1"}]',
+                method="POST",
+                auth=None,
+                ssl=False,
+            )
+
+    @pytest.mark.asyncio
+    async def test_set_heating_curve_gradient(self) -> None:
+        with aioresponses() as mock_keenergy_api:
+            mock_keenergy_api.post(
+                "http://mocked-host/var/readWriteVars?action=set",
+                payload={},
+                headers={"Content-Type": "application/json;charset=utf-8"},
+            )
+
+            client: KebaKeEnergyAPI = KebaKeEnergyAPI(host="mocked-host")
+            await client.heat_circuit.set_heating_curve_gradient(0.5)
+
+            mock_keenergy_api.assert_called_once_with(
+                url="http://mocked-host/var/readWriteVars?action=set",
+                data='[{"name": "APPL.CtrlAppl.sParam.heatCircuit[0].param.heatCurveGradient", "value": "0.5"}]',
+                method="POST",
+                auth=None,
+                ssl=False,
+            )
+
 
 class TestSolarCircuitSection:
     @pytest.mark.asyncio